add markdownToHtml Hooks
